backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.15.3
pyjwt==2.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
gunicorn==20.1.0
python-dotenv==0.21.0
